ai-video-summarizer/pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "flask>=3.1.2",
    "flask-cors>=6.0.2",
    "flask-dance>=7.1.0",
    "flask-login>=0.6.3",
    "flask-sqlalchemy>=3.1.1",
    "google-generativeai>=0.8.5",
    "gunicorn>=23.0.0",
    "nltk>=3.9.2",
    "oauthlib>=3.3.1",
    "openai>=2.11.0",
    "psycopg2-binary>=2.9.11",
    "pyjwt>=2.10.1",
    "python-dotenv>=1.2.1",
    "reportlab>=4.4.6",
    "requests>=2.32.5",
    "sqlalchemy>=2.0.45",
    "werkzeug>=3.1.4",
    "yt-dlp>=2025.12.8",
]
